refactor(edge-devices): move tables to view
