+++ app/edge/EdgeDevices/EdgeDevicesView/AMTDevicesDatatable/columns/index.tsx
