+++ app/edge/EdgeDevices/EdgeDevicesView/EdgeDevicesDatatable/columns/index.tsx
