+++ app/edge/EdgeDevices/EdgeDevicesView/index.ts
@@ -1,0 +1,1 @@
+export { EdgeDevicesView, EdgeDevicesViewAngular } from './EdgeDevicesView';
