--- app/edge/devices/components/EdgeDevicesDatatable/types.ts
@@ -1,4 +1,0 @@
-export interface Pagination {
-  pageLimit: number;
-  page: number;
-}
